requirements.txt
ibapi==9.81.1.post1
ib-insync==0.9.86
pandas==2.0.3
pytz==2024.1
SQLAlchemy==2.0.30
pandas_ta==0.3.14b0
numpy==1.24.4
matplotlib==3.7.5
seaborn==0.13.2
scikit-learn==1.3.2
joblib==1.4.2
holidays==0.49
python-dotenv==1.0.1
python-telegram-bot==21.2
requests==2.31.0
aiohttp==3.10.11